# backend/public/swagger.yaml
openapi: 3.0.0
info:
  title: Car Management API
  version: 1.0.0
  description: API documentation for the Car Management Application
servers:
  - url: http://localhost:5050
    description: Local server
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    User:
      type: object
      required:
        - username
        - password
      properties:
        username:
          type: string
          description: The user's unique username
        password:
          type: string
          description: The user's password
...
security:
  - bearerAuth: []
paths:
  /api/users/signup:
    post:
      summary: Register a new user
      tags:
        - User
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/User'
              example:
                username: 'john.doe@example.com'
                password: 'password123'
      responses:
        '201':
          description: User created successfully
        '400':
          description: Username already exists
        '500':
          description: Server error
  /api/users/login:
    post:
      summary: Log in a user
      tags:
        - User
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/User'
      responses:
        '200':
          description: Successful login
        '400':
          description: Invalid username or password
        '500':
          description: Server error
...